refactor :: entity 수정
